Personal/namalu/correct warnings and podspec file (#139)
* Corrected compiler warnings and fixed error in podspec file.

* updated version number in podspec

+++ HealthVault/Classes/Caching/Private/MHVCacheStatus.m
@@ -38,8 +38,8 @@
     {
         _lastCompletedSyncDate = [cachedRecord.lastSyncDate copy];
         _lastCacheConsistencyDate = [cachedRecord.lastConsistencyDate copy];
-        _newestCacheSequenceNumber = cachedRecord.newestCacheSequenceNumber;
-        _newestHealthVaultSequenceNumber = cachedRecord.newestHealthVaultSequenceNumber;
+        _newestCacheSequenceNumber = (NSInteger)cachedRecord.newestCacheSequenceNumber;
+        _newestHealthVaultSequenceNumber = (NSInteger)cachedRecord.newestHealthVaultSequenceNumber;
         _isCacheValid = cachedRecord.isValid;
     }
